backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.18
openai>=1.50.0
paddleocr>=3.0.0
paddlepaddle>=3.0.0
Pillow>=11.0.0
python-dotenv>=1.0.1
email-validator>=2.0.0
greenlet>=3.0.0
